~ progress on modernized provider

+++ src/ParaglidingWeather.Core/Types/IPrecipitation.cs
@@ -14,6 +14,6 @@ namespace ParaglidingWeather.Core.Types
         /// </summary>
         /// <param name="unit">The unit system to use.</param>
         /// <returns>Returns the value of precipitation.</returns>
-        int GetPrecipitation(Units.Precipitation unit);
+        double GetPrecipitation(Units.Precipitation unit);
     }
 }
